% probes file

\name{Probes.csv}
\docType{data}
\alias{Probes}

\non_function{}

\title{
  List of probes and probe-ID's.
}

\format{
  A \code{.csv}-file containing the probenames and probe-ID's of
  matches, mismatches, ... The file should contain at least 2 columns,
  with the first column named "Probename". The names of the next columns
  can be chosen, and could be something like "Matchprobe",
  "Mismatchprobe1" and "Mismachprobe2".
}

\description{
  }

\author{
  Kurt Sys (\email{kurt.sys@advalvas.be})
}

\seealso{
  \code{\link[base]{read.csv}}
}

\keyword{
  datasets
  probes
}
